feat: move request location permissions
