+++ src/modules/location/requestAndroidLocationPermissions.ts
@@ -1,0 +1,8 @@
+import { LocationManager } from "./LocationManager";
+
+/**
+ * @deprected Use `LocationManager.requestAndroidPermissions` instead
+ */
+export async function requestAndroidLocationPermissions(): Promise<boolean> {
+  return LocationManager.requestAndroidPermissions();
+}
